styles refactoring
