+++ src/App/App.js
@@ -1,18 +1,33 @@
 import React from 'react';
 import {Provider} from "react-redux";
 import store from "../redux/store";
-import Header from "../Header/Header";
+import ControlPanel from "../ControlPanel/ControlPanel";
 import ArrayView from "../ArrayView/ArrayView";
 import '@fortawesome/fontawesome-free/css/all.min.css';
 import 'bootstrap-css-only/css/bootstrap.min.css';
 import 'mdbreact/dist/css/mdb.css';
+import {MDBCol, MDBContainer, MDBRow} from "mdbreact";
+import SortingInfo from "../SortingInfo/SortingInfo";
 
 export default function App() {
   return (
     <Provider store={store}>
       <div className="App">
-        <Header />
-        <ArrayView />
+        <MDBContainer fluid={true}>
+          <MDBRow>
+            <MDBCol xl="9" lg="8">
+              <ArrayView />
+            </MDBCol>
+            <MDBCol xl="3" lg="4" className="d-none d-lg-block">
+              <SortingInfo />
+            </MDBCol>
+          </MDBRow>
+          <MDBRow>
+            <MDBCol lg="12">
+              <ControlPanel />
+            </MDBCol>
+          </MDBRow>
+        </MDBContainer>
       </div>
     </Provider>
   );
