+++ src/ArrayView/ArrayView.css
@@ -1,19 +1,8 @@
-.array-view {
-  display: flex;
-  flex-direction: row;
-}
-
 .array-container {
   width: 100%;
-  height: 80vh;
+  height: 65vh;
   display: flex;
   flex-direction: row;
   align-items: flex-end;
   justify-content: space-around;
-  flex: 3 1 0;
-}
-
-.sorting-info {
-  flex: 1 1 0;
-  align-self: center;
 }
