--- src/Header/Select/Select.css
@@ -1,4 +1,0 @@
-.select {
-  display: inline-block;
-  margin: 1em;
-}
