--- src/Header/Slider/Slider.css
@@ -1,4 +1,0 @@
-.range-slider {
-  display: inline-block;
-  margin: 1em;
-}
