+++ src/SortingInfo/SortingInfo.css
@@ -1,0 +1,17 @@
+.sorting-info-card {
+  border-radius: .25rem;
+  box-shadow: 0 2px 5px 0 rgba(0,0,0,0.16), 0 2px 10px 0 rgba(0,0,0,0.12);
+  margin: 1rem;
+  padding: 1rem;
+  height: 95%;
+}
+
+.sorting-info-card .header {
+  font-weight: bold;
+  margin-bottom: 3rem;
+}
+
+.sorting-info-card .info {
+  font-weight: bolder;
+  margin-right: .5rem;
+}
